Convert error_handling integration tests
- move content_store_throws_exception_for in content
  store helper (first use)
